Updating to Moodle up to version 4.5.
