+++ classes/text_filter.php
@@ -1,0 +1,129 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Moodle Ace inline content filter.
+ *
+ * @package    filter_ace_inline
+ * @copyright  2021 Richard Lobb
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace filter_ace_inline;
+
+use core\context;
+
+if (class_exists('\core_filters\text_filter')) {
+    class_alias('\core_filters\text_filter', 'filter_ace_inline_base_text_filter');
+} else {
+    class_alias('\moodle_text_filter', 'filter_ace_inline_base_text_filter');
+}
+
+/**
+ *
+ */
+class text_filter extends \filter_ace_inline_base_text_filter {
+
+    /**
+     * @var moodle_page page object.
+     */
+    protected $page;
+    /**
+     * @var context The context we are in.
+     */
+    protected $context;
+    /**
+     * @var array Any options for this filter in this context.
+     */
+    protected $options;
+
+    /**
+     * This function is called by the filter system to setup the filter.
+     *
+     * @param \moodle_page $page The page we are going to add requirements to.
+     * @param context $context The context which contents are going to be filtered.
+     * @return void
+     */
+    public function setup($page, $context) {
+        $this->page = $page;
+        $this->context = $context;
+        \qtype_coderunner_util::load_ace();
+    }
+
+    /**
+     * This function is called by the filter system to filter the text.
+     *
+     * @param string $text The text to filter.
+     * @param array $options The filter options.
+     * @return string The filtered text.
+     * @throws \dml_exception
+     */
+    public function filter($text, array $options = []) {
+        $this->options = $options;
+        // Basic test to avoid work.
+        if (!is_string($text)) {
+            // Non-string content can not be filtered anyway.
+            return $text;
+        }
+        $config = [
+            'button_label' => get_config('filter_ace_inline', 'button_label'),
+            'dark_theme_mode' => get_config('filter_ace_inline', 'dark_theme_mode'),
+        ];
+        $this->do_ace_highlight($text, $config);
+        $this->do_ace_interactive($text, $config);
+        return $text;
+    }
+
+    /**
+     * Process the given text by replacing any <pre> elements of class
+     * ace-highlight-code with an ace code high-lighted version.
+     * The actual work is done by JavaScript; this function just calls the
+     * appropriate function. The call to strpos is required regardless becuase
+     * apparently Mathjax generates a small content fragment, which is passed
+     * through all filters, on all content pages, even editing pages. We
+     * don't wish to use our filter on pages being edited.
+     * @param {string} $text The text to be processed.
+     * @param {array} $config The plugin configuration info.
+     * @return {string} The processed text.
+     */
+    public function do_ace_highlight($text, $config) {
+        if (strpos($text, 'ace-highlight-code') !== false) {
+            $this->page->requires->js_call_amd('filter_ace_inline/ace_inline_code',
+                'initAceHighlighting', [$config]);
+        }
+
+        return $text;
+    }
+
+    /**
+     * Process the given text by replacing any <pre> elements of class
+     * ace-interactive-code with an ace editor plus a Try it! button.
+     * The call to strpos is required regardless becuase
+     * apparently Mathjax generates a small content fragment, which is passed
+     * through all filters, on all content pages, even editing pages. We
+     * don't wish to use our filter on pages being edited.
+     * @param {string} $text The text to be processed.
+     * @param {array} $config The plugin configuration info.
+     * @return {string} The processed text
+     */
+    public function do_ace_interactive($text, $config) {
+        if (strpos($text, 'ace-interactive-code') !== false) {
+            $this->page->requires->js_call_amd('filter_ace_inline/ace_inline_code',
+                'initAceInteractive', [$config]);
+        }
+        return $text;
+    }
+}
