+++ filter.php
@@ -22,86 +22,4 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-defined('MOODLE_INTERNAL') || die();
-
-require_once($CFG->dirroot . '/question/type/coderunner/classes/util.php');
-
-/**
- * Class which implements the ace-inline filter.
- *
- * This filter looks for <pre> elements of class 'ace-highlight-code' or
- * 'ace-interactive-code' in Moodle question text and
- * replaces them an Ace editor panel to allow display and (for ace-interactive-code)
- * editing and execution, of program code.
- */
-class filter_ace_inline extends moodle_text_filter {
-
-    /**
-     * @var moodle_page page object.
-     */
-    protected $page;
-    protected $context;
-    protected $options;
-
-    public function setup($page, $context) {
-        $this->page = $page;
-        $this->context = $context;
-        qtype_coderunner_util::load_ace();
-    }
-
-    public function filter($text, array $options = []) {
-        $this->options = $options;
-        // Basic test to avoid work.
-        if (!is_string($text)) {
-            // Non-string content can not be filtered anyway.
-            return $text;
-        }
-        $config = array(
-            'button_label' => get_config('filter_ace_inline', 'button_label'),
-            'dark_theme_mode' => get_config('filter_ace_inline', 'dark_theme_mode')
-        );
-        $this->do_ace_highlight($text, $config);
-        $this->do_ace_interactive($text, $config);
-        return $text;
-    }
-
-    /**
-     * Process the given text by replacing any <pre> elements of class
-     * ace-highlight-code with an ace code high-lighted version.
-     * The actual work is done by JavaScript; this function just calls the
-     * appropriate function. The call to strpos is required regardless becuase
-     * apparently Mathjax generates a small content fragment, which is passed
-     * through all filters, on all content pages, even editing pages. We
-     * don't wish to use our filter on pages being edited.
-     * @param {string} $text The text to be processed.
-     * @param {array} $config The plugin configuration info.
-     * @return {string} The processed text.
-     */
-    public function do_ace_highlight($text, $config) {
-        if (strpos($text, 'ace-highlight-code') !== false) {
-            $this->page->requires->js_call_amd('filter_ace_inline/ace_inline_code',
-                    'initAceHighlighting', array($config));
-        }
-
-        return $text;
-    }
-
-    /**
-     * Process the given text by replacing any <pre> elements of class
-     * ace-interactive-code with an ace editor plus a Try it! button.
-     * The call to strpos is required regardless becuase
-     * apparently Mathjax generates a small content fragment, which is passed
-     * through all filters, on all content pages, even editing pages. We
-     * don't wish to use our filter on pages being edited.
-     * @param {string} $text The text to be processed.
-     * @param {array} $config The plugin configuration info.
-     * @return {string} The processed text
-     */
-    public function do_ace_interactive($text, $config) {
-        if (strpos($text, 'ace-interactive-code') !== false) {
-            $this->page->requires->js_call_amd('filter_ace_inline/ace_inline_code',
-                    'initAceInteractive', array($config));
-        }
-        return $text;
-    }
-}
+class_alias(\filter_ace_inline\text_filter::class, \filter_ace_inline::class);
