+++ lang/en/filter_ace_inline.php
@@ -13,6 +13,7 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
 /**
  * Strings for component 'filter_ace_inline', language 'en'
  *
@@ -20,6 +21,9 @@
  * @copyright  Richard Jones, Michelle Hsieh 2017, 2022
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+// phpcs:disable moodle.Files.LangFilesOrdering.UnexpectedComment
+// phpcs:disable moodle.Files.LangFilesOrdering.IncorrectOrder
 
 $string['filtername'] = 'Ace inline';
